feat: Add Management and Management Portal content types
Introduces two new Strapi content types to establish a resource portal for affiliates.

The 'Management Portal' is a single-type schema that serves as the main page for this new section.

The 'Management' collection-type allows for creating individual, downloadable resources like publications and guides, which can be related to the portal. Both new types support internationalization.